Fix race condition allowing lost events.
For the previous `Async.race (takeMVar excVar) getMsg`:

What if `takeMVar` and `getMsg` happen at the same time, but the thread in which
getMsg finished returns a tiny bit quicker?
Then we lose the result of the `takeMVar`.

We could do:

    fOrErr <- liftIO (Async.race (readMVar excVar) getMsg)
    case fOrErr of
      Left err -> do
        _ <- liftIO $ takeMVar excVar -- clear the MVar
        onErr (go compRoot rendered) rendered ECThread err

but this wouldn't guarantee that we lose no `getMsg` results (which uses
`readChan dispatchChan` inside, and for `Chan`s no blocking peek function
exists).
And losing them would be bad in case the exception handler asks us to ignore the
exception (`HEIgnore`).

So instead, this commit switches to a signle channel for incoming events
(which are either updates or exceptions), which can be consumed race-free.

This means exceptions can no longer "overtake" events.
But that is fine, becuause events are usually handled swiftly.