fix problem highlighted in #7, preserving original semantics
The fix in #7 removed the semantics whereby exceptions are handled
immediately. This fix keeps that semantics in place.

We resort to STM, which is probably slower than `MVar` and `Async.race`,
but on the other hand we do not use additional threads at all, which
might make up for the loss of performance... I don't think it matters
either way.